# docs/openapi.yaml
openapi: 3.0.0
info:
  title: GPC Reliability Tracker API
  description: API for tracking reliability projects, vendors, and alerts.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base API URL
paths:
...
  /alerts:
    get:
      summary: List alerts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: is_acknowledged
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
  /alerts/{id}/acknowledge:
    post:
      summary: Acknowledge an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert acknowledged
...
components:
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        severity:
          type: string
        is_acknowledged:
          type: boolean
        created_at:
          type: string